Adjusting interop tests for 5295
